refactor(satinfo): Satellite info into separated file
